Restructured code: separate templates and .js per tab
