--- demosaurus-webapp/demosaurus/auth.py
@@ -1,9 +1,0 @@
-from flask_login import LoginManager
-from demosaurus.models import UserModel
-
-login_manager = LoginManager()
-
-
-@login_manager.user_loader
-def load_user(id):
-    return UserModel.query.get(int(id))
