+++ demosaurus-webapp/demosaurus/templates/base.html
@@ -1,19 +1,20 @@
 <!doctype html>
-<link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}">
-<link rel="stylesheet" href="//code.jquery.com/ui/1.12.1/themes/base/jquery-ui.css">
+
+
+<!-- Load JQuery and DataTables -->
 <script type=text/javascript src="{{url_for('static', filename='js/jquery-3.4.1.js') }}"></script>
-<script type=text/javascript src="{{url_for('static', filename='js/annif.js') }}"></script>
-<script type=text/javascript src="{{url_for('static', filename='js/my_functions.js') }}"></script>
-<script type=text/javascript> $SCRIPT_ROOT = {{ request.script_root|tojson|safe }};</script>
 <script src="https://code.jquery.com/ui/1.12.1/jquery-ui.min.js"></script>
+<link rel="stylesheet" href="//code.jquery.com/ui/1.12.1/themes/base/jquery-ui.css">
 
-<script src="https://kit.fontawesome.com/abcca5fd9b.js" crossorigin="anonymous"></script>
-
-
-<link rel="stylesheet" type="text/css" href="https://cdn.datatables.net/1.11.0/css/jquery.dataTables.min.css"/>
- 
 <script type="text/javascript" charset="utf8" src="https://cdn.datatables.net/1.11.0/js/jquery.dataTables.min.js"></script>
+<link rel="stylesheet" type="text/css" href="https://cdn.datatables.net/1.11.0/css/jquery.dataTables.min.css"/>
 
+<!-- Load fontawesome (??) and set $SCRIPT_ROOT (??) -->
+<script src="https://kit.fontawesome.com/abcca5fd9b.js" crossorigin="anonymous"></script>
+<script type=text/javascript> $SCRIPT_ROOT = {{ request.script_root|tojson|safe }};</script>
+
+<link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}">
+<script type=text/javascript src="{{url_for('static', filename='js/my_functions.js') }}"></script>
 
 
 {{ JSGlue.include() }}
