--- demosaurus-webapp/demosaurus/templates/login.html
@@ -1,8 +1,0 @@
-<form action="" method ="POST">
-    <p>email <input type ="email" name ="email" /></p>
-    <p>password <input type ="password" name ="password" /></p>
-    <p> submit <input type ="submit" value ="Submit" /></p>
-</form>
- 
-<h3>Dont Have an account??</h3>
-<h3><a href ="{{url_for('register') }}">Register Here</a></h3>
