+++ demosaurus-webapp/demosaurus/templates/publication/contributors-tab.html
@@ -1,0 +1,65 @@
+<table id="contributortable" >
+      <thead>
+      <tr>
+        <th><!--i class="fas fa-trash-alt"></i--></th>
+        <th><!--i class="fas fa-arrow-up"></i--></th>
+        <th><!--i class="fas fa-arrow-down"></i--></th>
+        <th class="name_cell"> Naam </th>
+        <th class="name_cell"> Rol </th>
+        <th class="name_cell"> PPN </th>
+      </tr>
+      </thead>
+      <tbody style="max-height: 120px"/>
+      <!-- This is where contributors may be pre-filled and/ or manually added (client-side)-->
+      <tfoot>
+      <tr align="right">
+        <td><input type="button" title="Voeg naam toe" value="&#xf234" class="fas fa-user-plus" onclick="add_contributor_row()" /></td>
+      </tr>
+      </tfoot>
+    </table>
+
+<div style="width:100%;float:left;">
+
+  <div class='bottomcontent' id="contributors_content" style="display: block;">
+    <h2 id=thesaureer_title></h2>
+    <div id='placeholder'></div>
+    <table id="candidate_list" class="fixed_header">
+      <thead/>
+      <tbody class="zebra"/>
+      <tfoot/>
+    </table>
+  </div>
+</div>
+
+<!-- thesaureerblock sub-scores tooltip table.-->
+<div class="divTable" id="thesMatchHover" style="display: none;">
+  <div class="divTableHeading">
+    <div class="divTableRow">
+      <div class="divTableHead"></div>
+      <div class="divTableHead">Score</div>
+      <div class="divTableHead">Confidence</div>
+    </div>
+  </div>
+  <div class="divTableBody">
+    <div class="divTableRow">
+      <div class="divTableCell" id="tttb13"><b>Naam</b></div>
+      <div class="divTableCell" id="tttb14">&nbsp;</div>
+      <div class="divTableCell" id="tttb15">&nbsp;</div>
+    </div>
+    <!--      <div class="divTableRow">-->
+    <!--        <div class="divTableCell" id="tttb1"><b>Rol</b></div>-->
+    <!--        <div class="divTableCell" id="tttb2">&nbsp;</div>-->
+    <!--        <div class="divTableCell" id="tttb3">&nbsp;</div>-->
+    <!--      </div>-->
+    <div class="divTableRow">
+      <div class="divTableCell" id="tttb4"><b>Genre</b></div>
+      <div class="divTableCell" id="tttb5">&nbsp;</div>
+      <div class="divTableCell" id="tttb6">&nbsp;</div>
+    </div>
+    <div class="divTableRow">
+      <div class="divTableCell" id="tttb7"><b>JvU</b></div>
+      <div class="divTableCell" id="tttb8">&nbsp;</div>
+      <div class="divTableCell" id="tttb9">&nbsp;</div>
+    </div>
+  </div>
+</div>
