+++ demosaurus-webapp/demosaurus/templates/publication/export-tab.html
@@ -1,0 +1,7 @@
+Controleer KMC's hieronder, kopiëer en plak naar WinIBW
+    <div id="message"> Eventuele waarschuwingen </div>
+
+
+<!-- KMC block. -->
+<div class="bottomcontent" id="export_content" style="display: block;">
+</div>
