--- demosaurus-webapp/demosaurus/templates/register.html
@@ -1,9 +1,0 @@
-<form action="" method ="POST">
-    <p>email <input type ="email" name ="email" /></p>
-    <p>Username <input type ="text" name ="username" /></p>
-    <p>password <input type ="password" name ="password" /></p>
-    <p> submit <input type ="submit" value ="Submit" /></p>
-</form>
- 
-<h3>Already Have an Account?</h3><br>
-<h3><a href ="{{url_for('login')}}">Login Here</a></h3>
